python/chatbot/sample-project-voice-chatbot/scripts/requirements.txt
# Galileo SDK for creating protect stages
galileo>=1.0.0
galileo-core>=0.1.0
python-dotenv>=1.0.0
